MAINT: One script one accessor
