+++ dtoolkit/accessor/series/__init__.py
@@ -7,8 +7,8 @@ from dtoolkit.accessor.series.expand import expand  # noqa: F401
 from dtoolkit.accessor.series.filter_in import filter_in  # noqa: F401
 from dtoolkit.accessor.series.getattr import getattr  # noqa: F401
 from dtoolkit.accessor.series.groupby_index import groupby_index  # noqa: F401
-from dtoolkit.accessor.series.jenks import jenks_bin  # noqa: F401
-from dtoolkit.accessor.series.jenks import jenks_breaks  # noqa: F401
+from dtoolkit.accessor.series.jenks_bin import jenks_bin  # noqa: F401
+from dtoolkit.accessor.series.jenks_breaks import jenks_breaks  # noqa: F401
 from dtoolkit.accessor.series.len import len  # noqa: F401
 from dtoolkit.accessor.series.query import query  # noqa: F401
 from dtoolkit.accessor.series.to_set import to_set  # noqa: F401
